+++ lang/da/messages.php
@@ -18,7 +18,7 @@ return [
 
     'association_history.guest' => 'gæst',
     'association_history.at' => 'kl.',
-    'association_history.created' => 'oprettet',
+    'association_history.created' => 'Post oprettet',
     'association_history.edited' => 'redigerede',
     'association_history.from' => 'fra',
     'association_history.to' => 'til',