+++ lang/en/messages.php
@@ -18,7 +18,7 @@ return [
 
     'association_history.guest' => 'guest',
     'association_history.at' => 'at',
-    'association_history.created' => 'created',
+    'association_history.created' => 'Entry created',
     'association_history.edited' => 'edited',
     'association_history.from' => 'from',
     'association_history.to' => 'to',