+++ resources/views/components/association-history.blade.php
@@ -13,7 +13,9 @@
                     <dd>
                         <em>{{ $history->column_name }}</em>
                         @if (! $history->is_complex)
-                            @lang('ig-common::messages.association_history.from') <samp title="{{ $history->column_prev_value ?? '–' }}">{{ Str::limit($history->column_prev_value ?? '–', 20) }}</samp>
+                            @if ($history->column_prev_value !== null && $history->column_prev_value !== '')
+                                @lang('ig-common::messages.association_history.from') <samp title="{{ $history->column_prev_value }}">{{ Str::limit($history->column_prev_value, 20) }}</samp>
+                            @endif
                             @lang('ig-common::messages.association_history.to') <samp title="{{ $history->new_value ?? '–' }}">{{ Str::limit($history->new_value ?? '–', 20) }}</samp>
                         @endif
                     </dd>